requirements.txt
torch~=2.3.0 # see https://pytorch.org/get-started/locally/ when it gets incompatible with the CUDA version
transformers~=4.41.2
pillow~=10.4.0
diffusers~=0.29.2
safetensors~=0.4.2
numpy~=1.26.4
scikit-image~=0.23.2
scikit-learn~=1.4.1.post1
torchvision~=0.18.0
matplotlib~=3.9.0
open3d~=0.18.0
scipy~=1.12.0
